ref: Move ChapterExtensions to util.chapter package
Moved `ChapterExtensions.kt` and its test file to `eu.kanade.tachiyomi.util.chapter` package to improve cohesion.
Updated imports in all consuming files.
Verified with `ktfmtFormatMain` and `app:compileStandardDebugKotlin`.
