+++ app/src/main/java/eu/kanade/tachiyomi/ui/main/MainActivity.kt
@@ -48,7 +48,7 @@ import eu.kanade.tachiyomi.ui.main.states.SideNavMode
 import eu.kanade.tachiyomi.ui.reader.ReaderActivity
 import eu.kanade.tachiyomi.ui.security.SecureActivityDelegate
 import eu.kanade.tachiyomi.util.chapter.ChapterItemSort
-import eu.kanade.tachiyomi.util.isAvailable
+import eu.kanade.tachiyomi.util.chapter.isAvailable
 import eu.kanade.tachiyomi.util.view.setComposeContent
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
