+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaConstants.kt
@@ -6,7 +6,7 @@ import eu.kanade.tachiyomi.data.database.models.Manga
 import eu.kanade.tachiyomi.data.database.models.MergeType
 import eu.kanade.tachiyomi.data.database.models.SourceMergeManga
 import eu.kanade.tachiyomi.data.external.ExternalLink
-import eu.kanade.tachiyomi.util.MissingChapterHolder
+import eu.kanade.tachiyomi.util.chapter.MissingChapterHolder
 import kotlinx.collections.immutable.ImmutableSet
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.PersistentSet
